// web/src/api/models/TypesArtifactContent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesArtifactContent
 */
export interface TypesArtifactContent {
    /**
     * 
     * @type {string}
     * @memberof TypesArtifactContent
     */
    content: string;
    /**
     * 
     * @type {string}
     * @memberof TypesArtifactContent
     */
    dest: string;
}

/**
 * Check if a given object implements the TypesArtifactContent interface.
 */
export function instanceOfTypesArtifactContent(value: object): value is TypesArtifactContent {
    if (!('content' in value) || value['content'] === undefined) return false;
    if (!('dest' in value) || value['dest'] === undefined) return false;
    return true;
}

export function TypesArtifactContentFromJSON(json: any): TypesArtifactContent {
    return TypesArtifactContentFromJSONTyped(json, false);
}

export function TypesArtifactContentFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesArtifactContent {
    if (json == null) {
        return json;
    }
    return {
        
        'content': json['content'],
        'dest': json['dest'],
    };
}

export function TypesArtifactContentToJSON(json: any): TypesArtifactContent {
    return TypesArtifactContentToJSONTyped(json, false);
}

export function TypesArtifactContentToJSONTyped(value?: TypesArtifactContent | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'content': value['content'],
        'dest': value['dest'],
    };
}