// web/src/api/models/TypesFailureRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TypesArtifactError } from './TypesArtifactError';
import {
    TypesArtifactErrorFromJSON,
    TypesArtifactErrorFromJSONTyped,
    TypesArtifactErrorToJSON,
    TypesArtifactErrorToJSONTyped,
} from './TypesArtifactError';

/**
 * 
 * @export
 * @interface TypesFailureRecord
 */
export interface TypesFailureRecord {
    /**
     * 
     * @type {Array<string>}
     * @memberof TypesFailureRecord
     */
    artifacts?: Array<string>;
    /**
     * 
     * @type {Array<TypesArtifactError>}
     * @memberof TypesFailureRecord
     */
    errors?: Array<TypesArtifactError>;
    /**
     * 
     * @type {string}
     * @memberof TypesFailureRecord
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFailureRecord
     */
    snap_id?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFailureRecord
     */
    source?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFailureRecord
     */
    time: string;
}

/**
 * Check if a given object implements the TypesFailureRecord interface.
 */
export function instanceOfTypesFailureRecord(value: object): value is TypesFailureRecord {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('time' in value) || value['time'] === undefined) return false;
    return true;
}

export function TypesFailureRecordFromJSON(json: any): TypesFailureRecord {
    return TypesFailureRecordFromJSONTyped(json, false);
}

export function TypesFailureRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesFailureRecord {
    if (json == null) {
        return json;
    }
    return {
        
        'artifacts': json['artifacts'] == null ? undefined : json['artifacts'],
        'errors': json['errors'] == null ? undefined : ((json['errors'] as Array<any>).map(TypesArtifactErrorFromJSON)),
        'id': json['id'],
        'snap_id': json['snap_id'] == null ? undefined : json['snap_id'],
        'source': json['source'] == null ? undefined : json['source'],
        'time': json['time'],
    };
}

export function TypesFailureRecordToJSON(json: any): TypesFailureRecord {
    return TypesFailureRecordToJSONTyped(json, false);
}

export function TypesFailureRecordToJSONTyped(value?: TypesFailureRecord | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'artifacts': value['artifacts'],
        'errors': value['errors'] == null ? undefined : ((value['errors'] as Array<any>).map(TypesArtifactErrorToJSON)),
        'id': value['id'],
        'snap_id': value['snap_id'],
        'source': value['source'],
        'time': value['time'],
    };
}